django_smoke_app/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
pip==22.3.1
setuptools==65.5.1
sqlparse==0.4.3
wheel==0.38.4
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pymongo
dnspython3
djangorestframework==3.14.0